+++ CHANGELOG.md
@@ -4,4 +4,10 @@ Change Log
 2017-01-22
 ----------
 
+ * addition of pop, shift, slice and each
+ * fixed Immutable
+
+2017-01-22
+----------
+
 Initial commit, forked from domain input mapper.